feat: add JWT and AppRole login support, enhance secret management
* Introduced new packages for JWT source handling and AppRole login functionality.
* Implemented secret path building and extraction methods for improved secret management.
* Added client interface for reading and writing secrets, enhancing interaction with Vault.
* Updated provider configuration to include JWT authentication methods.

These changes expand the plugin's capabilities for authentication and secret retrieval, improving flexibility and usability.

updated dependencies to use new apis
