+++ internal/vaultcompat/jwtsource/source.go
@@ -2,7 +2,6 @@ package jwtsource
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"os"
 	"strings"
@@ -40,29 +39,4 @@ func (f File) Token(context.Context) (string, error) {
 	}
 
 	return token, nil
-}
-
-type Chain struct {
-	Sources []Source
-}
-
-func (c Chain) Token(ctx context.Context) (string, error) {
-	var errs []error
-	for _, source := range c.Sources {
-		if source == nil {
-			continue
-		}
-
-		token, err := source.Token(ctx)
-		if err == nil {
-			return token, nil
-		}
-		errs = append(errs, err)
-	}
-
-	if len(errs) == 0 {
-		return "", fmt.Errorf("no jwt sources configured")
-	}
-
-	return "", errors.Join(errs...)
 }
